broker/requirements.txt
kafka-python==2.1.2
praw==7.8.1
python-dotenv==1.0.1
requests==2.31.0 